+++ packages/editor/editor-core/examples/16-Mobile.tsx
@@ -2,11 +2,7 @@ import * as React from 'react';
 import { Editor } from '../src/index';
 import EditorContext from '../src/ui/EditorContext';
 import WithEditorActions from '../src/ui/WithEditorActions';
-import {
-  MentionDescription,
-  MentionProvider,
-  MentionResourceConfig,
-} from '@atlaskit/mention';
+import { MentionDescription, MentionProvider } from '@atlaskit/mention';
 
 class MentionProviderImpl implements MentionProvider {
   filter(query?: string): void {}
@@ -17,19 +13,8 @@ class MentionProviderImpl implements MentionProvider {
   isFiltering(query: string): boolean {
     return false;
   }
-  subscribe(
-    key: string,
-    callback?,
-    errCallback?,
-    infoCallback?,
-    allResultsCallback?,
-  ): void {}
+  subscribe(): void {}
   unsubscribe(key: string): void {}
-  getMentionConfig(): MentionResourceConfig {
-    return {
-      url: 'https://example.com/',
-    };
-  }
 }
 
 export function mobileEditor() {